added: opm/models/utils/terminal.[ch]pp
holds some terminal related methods that used to sit
in parametersystem
